feat: add shell application (#20)
<!-- Please provide enough information so that others can review your
pull request. -->
<!-- Keep pull requests small and focused on a single change. -->

### Summary
Added Shell application. Shell is a blueprint of Host application with
simple UI and shared dependencies to work exactly the same as the Host
application with dependencies. It could be provided for dev team to test
exposed mini apps as a part of Host app.
<!-- What existing problem does the pull request solve? Can you solve
the issue with a different approach? -->

### Test plan
1. Run `start:standalone:shell`
2. Run other mini apps bundlers
3. Run Shell app
Open Shell app
<!-- List the steps with which we can test this change. Provide
screenshots if this changes anything visual. -->