fix: don't set ScriptManager storage in dev (#101)
### Summary

A faulty hot update can be incorrectly cached by the `ScriptManager`,
e.g. making syntax error in a JSON file, resulting in error persisting
after reloads, when cache is configured to be stored in AsyncStorage.
 
Temporary solve this by not applying cache for dev. Left a todo to
address that properly.